add integration-test addRoute

+++ backend/src/main/java/com/github/astridottmann/backend/models/Car.java
@@ -10,7 +10,7 @@ public class Car extends Vehicle {
     private String fuel;
     private String carSize;
 
- public Car(String type, float co2Emission, String fuel, String carSize) {
+ public Car(String type, double co2Emission, String fuel, String carSize) {
         super(type, co2Emission);
         this.fuel = fuel;
         this.carSize = carSize;
